IMPALA-3656: Hitting DCHECK/CHECK does not write minidumps
When hitting a DCHECK/CHECK the daemons do not write minidumps. This
is caused by glog's own stack unwinding mechanism, which catches
SIGABRT and removes all other handlers before aborting.

This patch backports a change from glog, which only resets the
SIGABRT handler, if it is the one installed by glog itself.

https://github.com/google/glog/commit/cda16b3443e2d6ef88cdbbe10b9a11adea6f33fe

Testing:
Did a local build and made sure that both the DCHECK error are logged
and minidumps are written.