Fix setup / download order in build scripts
We need to download package sources before setting up the build
environment for package builds. The naming of the 'header' function made
it sound like it could be called before downloading the sources. This
change fixes the places where the order was wrong and renames the header
function to setup_package_build to make its name less confusing.

Change-Id: I287f091c8e772c08c8a61644d08117145b1ef16d